pyproject.toml
[tool.poetry]
name = "presidio-pii-masking"
version = "0.1.0"
description = "PII（個人情報）検出・匿名化ツール（Microsoft Presidioベース）"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "presidio_pii_masking", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
presidio-analyzer = "^2.2"
presidio-anonymizer = "^2.2"
spacy = "^3.7"
pydantic = "^2.4"
typer = "^0.9"
rich = "^13.4"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3"
black = "^23.3"
ruff = "^0.0.270"
mypy = "^1.3"
isort = "^5.12"
pytest-cov = "^6.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
pii-mask = "presidio_pii_masking.cli:app"

[tool.ruff]
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "I",  # isort
    "W",  # pycodestyle warnings
]
ignore = []
line-length = 100

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_ignores = true 